// backend/src/modules/production-orders/dto/create-production-order-leak-tightness-check.dto.ts
export class CreateProductionOrderLeakTightnessCheckDto {
  requirement?: string | null;
  unit_1_result?: boolean | number | string | null;
  unit_2_result?: boolean | number | string | null;
  unit_3_result?: boolean | number | string | null;
  unit_4_result?: boolean | number | string | null;
  unit_5_result?: boolean | number | string | null;
  unit_6_result?: boolean | number | string | null;
  unit_7_result?: boolean | number | string | null;
  unit_8_result?: boolean | number | string | null;
  unit_9_result?: boolean | number | string | null;
  unit_10_result?: boolean | number | string | null;
}